requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
pandas==2.1.3
beautifulsoup4==4.12.2
selenium==4.15.0
webdriver-manager==4.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.7
